Split views in separate files
